backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn
pymongo
requests
python-dotenv
openai
google-cloud-aiplatform
